3included/opendata-to-db/requirements.txt
requests
mysql-connector-python
pandas
openpyxl